Checking that lpReserved is NULL before attempting clean-up. (fixes: http://easyhook.codeplex.com/workitem/12795)
As per: http://msdn.microsoft.com/en-us/library/windows/desktop/ms682583(v=vs.85).aspx
...
When handling DLL_PROCESS_DETACH, a DLL should free resources such as heap memory only if the DLL is being unloaded dynamically (the lpReserved parameter is NULL). If the process is terminating (the lpvReserved parameter is non-NULL), all threads in the process except the current thread either have exited already or have been explicitly terminated by a call to the ExitProcess function, which might leave some process resources such as heaps in an inconsistent state. In this case, it is not safe for the DLL to clean up the resources. Instead, the DLL should allow the operating system to reclaim the memory.
...

git-svn-id: https://easyhook.svn.codeplex.com/svn@66701 142e9c35-602f-4880-9f37-ec48a24daa6a